requirements.txt
cffi==1.17.1
diffusers==0.30.2
einops==0.8.1
h5py==3.11.0
imageio==2.34.0
matplotlib==3.8.3
numpy==2.2.4
omegaconf==2.3.0
opencv_contrib_python==4.10.0.84
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
pandas==1.5.3
Pillow==11.2.1
scipy==1.15.2
skimage==0.0
tabulate==0.9.0
tensorboardX==2.2
tensorboardX==2.6.2.2
timm==0.5.4
tqdm==4.65.2
transformers==4.47.1
wandb==0.19.9
